updated .gitignore and removed commented out lines from main

+++ src/airport/FoodCourt.java
@@ -3,9 +3,9 @@ package airport;
 public class FoodCourt {
     public static String foodType;
     public static String restaurantName;
-    public static float price;
+    public static double price;
 
-    public FoodCourt(String food, String place, float fPrice) {
+    public FoodCourt(String food, String place, double fPrice) {
         this.foodType = food;
         this.restaurantName = place;
         this.price = fPrice;
@@ -19,7 +19,7 @@ public class FoodCourt {
         return restaurantName;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
